Move all specs to the src directory
