+++ .github/whitelist-capital-letters.txt
@@ -25,14 +25,6 @@
 ./libs/angular/src/services/theming/themeBuilder.ts
 ./libs/angular/src/interfaces/selectOptions.ts
 ./libs/components/src/stories/Introduction.stories.mdx
-./libs/common/spec/web/services/webCryptoFunction.service.spec.ts
-./libs/common/spec/shared/interceptConsole.ts
-./libs/common/spec/models/domain/symmetricCryptoKey.spec.ts
-./libs/common/spec/models/domain/encArrayBuffer.spec.ts
-./libs/common/spec/matchers/toEqualBuffer.spec.ts
-./libs/common/spec/matchers/toEqualBuffer.ts
-./libs/common/spec/services/stateMigration.service.spec.ts
-./libs/common/spec/services/consoleLog.service.spec.ts
 ./libs/common/src/misc/nodeUtils.ts
 ./libs/common/src/misc/linkedFieldOption.decorator.ts
 ./libs/common/src/misc/serviceUtils.ts
