--- apps/cli/spec/bw.spec.ts
@@ -1,3 +1,0 @@
-describe("bw", () => {
-  test.todo("is a placeholder test");
-});
