+++ libs/common/spec/matchers/index.ts
@@ -1,0 +1,1 @@
+export * from "./to-equal-buffer";
