+++ libs/common/src/services/encrypt.service.spec.ts
@@ -1,15 +1,15 @@
 import { mockReset, mock } from "jest-mock-extended";
 
-import { CryptoFunctionService } from "@bitwarden/common/abstractions/cryptoFunction.service";
-import { LogService } from "@bitwarden/common/abstractions/log.service";
-import { EncryptionType } from "@bitwarden/common/enums";
-import { EncArrayBuffer } from "@bitwarden/common/models/domain/enc-array-buffer";
-import { EncString } from "@bitwarden/common/models/domain/enc-string";
-import { SymmetricCryptoKey } from "@bitwarden/common/models/domain/symmetric-crypto-key";
-import { EncryptServiceImplementation } from "@bitwarden/common/services/cryptography/encrypt.service.implementation";
+import { makeStaticByteArray } from "../../spec";
+import { CryptoFunctionService } from "../abstractions/cryptoFunction.service";
+import { LogService } from "../abstractions/log.service";
+import { EncryptionType } from "../enums";
+import { EncArrayBuffer } from "../models/domain/enc-array-buffer";
+import { EncString } from "../models/domain/enc-string";
+import { SymmetricCryptoKey } from "../models/domain/symmetric-crypto-key";
+import { CsprngArray } from "../types/csprng";
 
-import { CsprngArray } from "../../src/types/csprng";
-import { makeStaticByteArray } from "../utils";
+import { EncryptServiceImplementation } from "./cryptography/encrypt.service.implementation";
 
 describe("EncryptService", () => {
   const cryptoFunctionService = mock<CryptoFunctionService>();
