+++ libs/common/src/tools/send/models/domain/send-file.spec.ts
@@ -1,4 +1,4 @@
-import { mockEnc } from "../../../../../spec/utils";
+import { mockEnc } from "../../../../../spec";
 import { SendFileData } from "../data/send-file.data";
 
 import { SendFile } from "./send-file";
