+++ libs/common/src/vault/models/domain/cipher.spec.ts
@@ -2,7 +2,7 @@
 import { Substitute, Arg } from "@fluffy-spoon/substitute";
 import { Jsonify } from "type-fest";
 
-import { mockEnc, mockFromJson } from "../../../../spec/utils";
+import { mockEnc, mockFromJson } from "../../../../spec";
 import { FieldType, SecureNoteType, UriMatchType } from "../../../enums";
 import { EncString } from "../../../models/domain/enc-string";
 import { InitializerKey } from "../../../services/cryptography/initializer-key";
