+++ libs/common/src/vault/models/domain/login.spec.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line no-restricted-imports
 import { Substitute, Arg } from "@fluffy-spoon/substitute";
 
-import { mockEnc, mockFromJson } from "../../../../spec/utils";
+import { mockEnc, mockFromJson } from "../../../../spec";
 import { UriMatchType } from "../../../enums";
 import { EncString } from "../../../models/domain/enc-string";
 import { LoginData } from "../../models/data/login.data";
